Issue #11302: Converting to kernel test for better code coverage
