+++ profile/modules/os/modules/os_theme_preview/tests/src/ExistingSite/HelperTest.php
@@ -2,8 +2,8 @@
 
 namespace Drupal\Tests\os_theme_preview\ExistingSite;
 
+use Drupal\os_theme_preview\Helper;
 use Drupal\os_theme_preview\ThemePreviewException;
-use weitzman\DrupalTestTraits\ExistingSiteBase;
 
 /**
  * HelperTest.
@@ -12,22 +12,7 @@ use weitzman\DrupalTestTraits\ExistingSiteBase;
  * @group other
  * @coversDefaultClass \Drupal\os_theme_preview\Helper
  */
-class HelperTest extends ExistingSiteBase {
-
-  /**
-   * Helper Service.
-   *
-   * @var \Drupal\os_theme_preview\HelperInterface
-   */
-  protected $helper;
-
-  /**
-   * {@inheritdoc}
-   */
-  public function setUp() {
-    parent::setUp();
-    $this->helper = $this->container->get('os_theme_preview.helper');
-  }
+class HelperTest extends TestBase {
 
   /**
    * Negative test for startPreviewMode.
@@ -36,9 +21,25 @@ class HelperTest extends ExistingSiteBase {
    *
    * @throws \Drupal\os_theme_preview\ThemePreviewException
    */
-  public function testStartPreviewMode() {
+  public function testFalseStartPreviewMode() {
     $this->expectException(ThemePreviewException::class);
     $this->helper->startPreviewMode('hwpi_themeone_bentley');
+  }
+
+  /**
+   * Positive test for startPreviewMode.
+   *
+   * @covers ::startPreviewMode
+   *
+   * @throws \Drupal\os_theme_preview\ThemePreviewException
+   */
+  public function testTrueStartPreviewMode() {
+    $this->setSession($this->requestStack->getCurrentRequest());
+
+    $this->helper->startPreviewMode('hwpi_themeone_bentley');
+
+    $previewed_theme = $this->requestStack->getCurrentRequest()->getSession()->get(Helper::SESSION_KEY);
+    $this->assertEquals('hwpi_themeone_bentley', $previewed_theme);
   }
 
   /**
@@ -58,9 +59,26 @@ class HelperTest extends ExistingSiteBase {
    *
    * @throws \Drupal\os_theme_preview\ThemePreviewException
    */
-  public function testStopPreviewMode() {
+  public function testFalseStopPreviewMode() {
     $this->expectException(ThemePreviewException::class);
     $this->helper->stopPreviewMode();
+  }
+
+  /**
+   * Positive test for stopPreviewMode.
+   *
+   * @covers ::stopPreviewMode
+   *
+   * @throws \Drupal\os_theme_preview\ThemePreviewException
+   */
+  public function testTrueStopPreviewMode() {
+    $this->setSession($this->requestStack->getCurrentRequest());
+
+    $this->helper->startPreviewMode('hwpi_themeone_bentley');
+    $this->helper->stopPreviewMode();
+
+    $previewed_theme = $this->requestStack->getCurrentRequest()->getSession()->get(Helper::SESSION_KEY);
+    $this->assertNull($previewed_theme);
   }
 
 }
